+++ app/controllers/concerns/caluculater/base.rb
@@ -1,0 +1,10 @@
+module Caluculater
+  module Base
+    extend ActiveSupport::Concern
+
+    included do
+      include Caluculater::Common
+      include Caluculater::Mock
+    end
+  end
+end
